requirements.txt
psutil
pywinctl
pynput
google-generativeai
pyautogui
pytesseract
Pillow
pyaudio
pydub
speechrecognition
pyscreenshot
Pillow
tk
numpy
pyaudio
rumps
